Hide blended tab selector
Hide the tab selector in the basic search *if* TC article_discovery
tab is selected.

NOTE: If we want to hide the tab selector in all cases, regardless
of which tab/scope we're on, then no custom JS is needed. We can
just use CSS: `div.search-options { display: none !important; }`