Open ILLiad links in new tab
Due to the fact that we're routing ILLiad links through EZProxy
for Shib auth, using the back button to return to the Primo Account
page from ILLiad can result in a not-so-great experience. Opening
the links in a new window seems like the most prudent thing to do
in this case.